+++ src/test/validation_tests.cpp
@@ -404,6 +404,7 @@ BOOST_AUTO_TEST_CASE(test_mainnet_assumeutxo_snapshot_metadata)
         85'850,
         105'550,
         106'875,
+        118'225,
     };
 
     BOOST_REQUIRE_EQUAL(snapshot_heights.size(), expected_snapshot_heights.size());
@@ -420,6 +421,7 @@ BOOST_AUTO_TEST_CASE(test_mainnet_assumeutxo_snapshot_metadata)
     BOOST_CHECK(params->AssumeutxoForHeight(85'850));
     BOOST_CHECK(params->AssumeutxoForHeight(105'550));
     BOOST_CHECK(params->AssumeutxoForHeight(106'875));
+    BOOST_CHECK(params->AssumeutxoForHeight(118'225));
     BOOST_CHECK(!params->AssumeutxoForHeight(50'000));
     BOOST_CHECK(!params->AssumeutxoForHeight(0));
 }